+++ python_modules/matiPatch.py
@@ -3,49 +3,45 @@ import os
 import sys
 import shutil
 
-def patch(argv):
-	full = False
+def error(msg):
+    print('\x1b[6;31;28m' + str(msg) + '\x1b[0m')
 
-	for i in range(0, len(argv)):
-		if (argv[i] == "-f" or argv[i] == "-F"):
-			del argv[i]
-			full = True
-			break
+def patch(argv, full=False, dev=False):
 
-	if (len(argv) > 2):
-		print("\nUnexpected/Unknown arguments recieved!\n")
+	if (len(argv) == 2 and (len(argv[1]) < 2 or len(argv[1]) > 15)):
+		error("\nProject name must be 2-15 characters long\n")
 		exit(1)
 
 	current_folder = os.path.split(os.getcwd())[1]
 
 	if (len(argv) == 2 and argv[1] != current_folder and os.path.isdir("../" + argv[1])):
-		print("\nError, there is another folder with that name in the parent directory!\n")
+		error("\nError, there is another folder with that name in the parent directory!\n")
 		exit(1)
 
-	print ("\nNativeScript MACANA Angular-Template patcher")
+	print ("\nNativeScript MCN Angular-Template patcher")
 
 	try:
 		print("\nChecking for necesary dependencies...", end="")
 		if not (os.path.isdir("app")):
-			print("Error!\nCouldn't find 'app' folder!\n")
+			error("Error!\nCouldn't find 'app' folder!\n")
 			exit(1)
 		if not (os.path.isfile("package.json")):
-			print("Error!\nCouldn't find 'package.json' file!\n")
+			error("Error!\nCouldn't find 'package.json' file!\n")
 			exit(1)
 		if not (os.path.isfile("tsconfig.json")):
-			print("Error!\nCouldn't find 'tsconfig.json' file!\n")
+			error("Error!\nCouldn't find 'tsconfig.json' file!\n")
 			exit(1)
 		if not (os.path.isfile("app/package.json")):
-			print("Error!\nCouldn't find 'app/package.json' file!\n")
+			error("Error!\nCouldn't find 'app/package.json' file!\n")
 			exit(1)
 		if not (os.path.isfile("app/App_Resources/Android/app.gradle")):
-			print("Error!\nCouldn't find 'app/App_Resources/Android/app.gradle' file!\n")
+			error("Error!\nCouldn't find 'app/App_Resources/Android/app.gradle' file!\n")
 			exit(1)
 		print("Done!")
 
 		if (os.path.isdir("platforms") and full):
 			print("Deleting platforms folder...", end="")
-			shutil.rmtree("platforms")
+			if not dev: shutil.rmtree("platforms")
 			print("Done!")
 		if (os.path.isdir("node_modules") and full):
 			print("Deleting node_modules folder...", end="")
@@ -58,9 +54,10 @@ def patch(argv):
 		if not (len(argv) == 1  or (len(argv) == 2 and argv[1] == name)):
 			print("Project's name: " + argv[1] + "\n")
 			print("Attempting to change current folder's name to '" + argv[1] + "'...", end="")
-			os.chdir("..")
-			os.rename(name, argv[1])
-			os.chdir(argv[1])
+			if not dev:
+				os.chdir("..")
+				os.rename(name, argv[1])
+				os.chdir(argv[1])
 			name = argv[1]
 			print("Done!")
 		else:
@@ -75,9 +72,10 @@ def patch(argv):
 				if (result):
 					data[i] = result.group(1) + "org.nativescript." + name + result.group(2) + "\n" #re.match trims the line break
 					break
-		with open("package.json", 'w') as f:
-			for ln in data:
-				f.write(ln)
+		if not dev:
+			with open("package.json", 'w') as f:
+				for ln in data:
+					f.write(ln)
 		print("Done!")
 
 		# app/package.json
@@ -89,9 +87,10 @@ def patch(argv):
 				if (result):
 					data[i] = result.group(1) + "tns-" + name.lower() + "-ng" + result.group(2) + "\n" #re.match trims the line break
 					break
-		with open("app/package.json", 'w') as f:
-			for ln in data:
-				f.write(ln)
+		if not dev:
+			with open("app/package.json", 'w') as f:
+				for ln in data:
+					f.write(ln)
 		print("Done!")
 
 		# app/App_Resources/Android/app.gradle
@@ -103,9 +102,10 @@ def patch(argv):
 				if (result):
 					data[i] = result.group(1) + "org.nativescript." + name + result.group(2) + "\n" #re.match trims the line break
 					break
-		with open("app/App_Resources/Android/app.gradle", 'w') as f:
-			for ln in data:
-				f.write(ln)
+		if not dev:
+			with open("app/App_Resources/Android/app.gradle", 'w') as f:
+				for ln in data:
+					f.write(ln)
 		print("Done!")
 
 		try:
@@ -114,12 +114,12 @@ def patch(argv):
 			if (full or not os.path.isdir("platforms")):
 				print("------------------------------------------------------")
 				print("Adding android platform\n")
-				os.system("tns platform add android")
+				if not dev: os.system("tns platform add android")
 				print("------------------------------------------------------")
 			if (full or not os.path.isdir("platforms")):
 				print("------------------------------------------------------")
 				print("Installing node_modules\n")
-				os.system("tns install")
+				if not dev: os.system("tns install")
 				print("------------------------------------------------------")
 		except:
 			print("\nERROR! Make sure 'nodejs', 'npm' and 'tns' are UP TO DATE!\n")
@@ -134,31 +134,35 @@ def patch(argv):
 				if (result):
 					data[i] = result.group(1) + name + result.group(2) + "\n" #re.match trims the line break
 					break
-		with open("platforms/android/settings.gradle", 'w') as f:
-			for ln in data:
-				f.write(ln)
+		if not dev:
+			with open("platforms/android/settings.gradle", 'w') as f:
+				for ln in data:
+					f.write(ln)
 		print("Done!")
 
 		print("Modyfing platforms/android/app/build.gradle...", end="")
 		with open("platforms/android/app/build.gradle", 'r') as f:
 			filedata = f.read()
 		if not (os.path.isfile("platforms/android/app/build.gradel.bk")):
-			with open("platforms/android/app/build.gradle.bk", 'w') as f:
-				f.write(filedata)
+			if not dev:
+				with open("platforms/android/app/build.gradle.bk", 'w') as f:
+					f.write(filedata)
 		#filedata = filedata.replace('compile "', 'implementation "')
 		filedata = filedata.replace('debugCompile "', 'debugImplementation "')
-		with open("platforms/android/app/build.gradle", 'w') as f:
-			f.write(filedata)
+		if not dev:
+			with open("platforms/android/app/build.gradle", 'w') as f:
+				f.write(filedata)
 		print("Done!")
 
 		print("Congiguring '.gitignore'...", end="")
-		with open(".gitignore", 'w') as f:
-			f.write("platforms\n")
-			f.write("node_modules\n")
-			f.write("app/**/*.js")
+		if not dev:
+			with open(".gitignore", 'w') as f:
+				f.write("platforms\n")
+				f.write("node_modules\n")
+				f.write("app/**/*.js")
 		print("Done!")
 
 		print("\nEverything Done! All credits to our Lord and Saviour: Mati Ratcliffe\n")
 
 	except Exception as e:
-		print("Error!\n\n" + str(e) + "\n")
+		error("Error!\n\n" + str(e) + "\n")
